rag-project/requirements.txt
sentence-transformers==2.6.1
faiss-cpu==1.13.2
numpy==1.26.4